Configuration File Reading & Applying
We can now read a configuration file, and apply the settings from it.
There's no error handling for the configuration types really, so it will
crash if something's not right. The wiki should be updated to have a
tutorial on how to create configuration files.
We can also now reset to the default configuration file from the debug
menu, and there's placeholder UI for updating the configuration file/the
entire app.